refactor(cli): split materialized_fixture_bytes into SRP submodules
Extract each MaterializeKind handler into its own focused module under
crates/uselesskey-cli/src/materialize/, leaving the dispatcher as a thin
shell. Keeps byte-for-byte output identical and preserves the public
surface; no behavior changes.
