+++ crates/uselesskey-cli/src/lib.rs
@@ -15,10 +15,9 @@ use base64::Engine as _;
 use base64::engine::general_purpose::STANDARD as BASE64_STD;
 use serde::{Deserialize, Serialize};
 use thiserror::Error;
-use uselesskey_core::{Factory, Seed};
-#[cfg(feature = "rsa-materialize")]
-use uselesskey_rsa::{RsaFactoryExt, RsaSpec};
-use uselesskey_token::{TokenFactoryExt, TokenSpec};
+
+mod materialize;
+use materialize::materialized_fixture_bytes;
 
 /// Bundle manifest describing generated artifacts and handoff metadata.
 #[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
@@ -419,99 +418,6 @@ fn resolve_fixture_path(out_dir: &Path, target: &Path) -> PathBuf {
     }
 }
 
-fn materialized_fixture_bytes(spec: &MaterializeFixtureSpec) -> Result<Vec<u8>, MaterializeError> {
-    let label = spec
-        .label
-        .clone()
-        .unwrap_or_else(|| fallback_label(&spec.out));
-    let fx = Factory::deterministic_from_str(&spec.seed);
-
-    match spec.kind {
-        MaterializeKind::EntropyBytes => {
-            let len = spec.len.unwrap_or(32);
-            let seed = Seed::from_text(&spec.seed);
-            let mut bytes = vec![0u8; len];
-            seed.fill_bytes(&mut bytes);
-            Ok(bytes)
-        }
-        MaterializeKind::TokenJwtShape => Ok(fx
-            .token(&label, TokenSpec::oauth_access_token())
-            .value()
-            .as_bytes()
-            .to_vec()),
-        MaterializeKind::RsaPkcs8Der => {
-            #[cfg(feature = "rsa-materialize")]
-            {
-                Ok(fx
-                    .rsa(&label, RsaSpec::rs256())
-                    .private_key_pkcs8_der()
-                    .to_vec())
-            }
-            #[cfg(not(feature = "rsa-materialize"))]
-            {
-                Err(MaterializeError::InvalidManifest(
-                    "rsa.pkcs8_der requires uselesskey-cli feature `rsa-materialize`".to_string(),
-                ))
-            }
-        }
-        MaterializeKind::RsaPkcs8Pem => {
-            #[cfg(feature = "rsa-materialize")]
-            {
-                Ok(fx
-                    .rsa(&label, RsaSpec::rs256())
-                    .private_key_pkcs8_pem()
-                    .as_bytes()
-                    .to_vec())
-            }
-            #[cfg(not(feature = "rsa-materialize"))]
-            {
-                Err(MaterializeError::InvalidManifest(
-                    "rsa.pkcs8_pem requires uselesskey-cli feature `rsa-materialize`".to_string(),
-                ))
-            }
-        }
-        MaterializeKind::PemBlockShape => {
-            let len = spec.len.unwrap_or(256);
-            let seed = Seed::from_text(&spec.seed);
-            let mut bytes = vec![0u8; len];
-            seed.fill_bytes(&mut bytes);
-            let payload = BASE64_STD.encode(bytes);
-            let block_label = normalize_pem_label(&label);
-            let mut out = String::new();
-            let _ = writeln!(&mut out, "-----BEGIN {block_label}-----");
-            for chunk in payload.as_bytes().chunks(64) {
-                let _ = writeln!(
-                    &mut out,
-                    "{}",
-                    std::str::from_utf8(chunk).map_err(|err| {
-                        MaterializeError::InvalidManifest(format!(
-                            "generated base64 payload was not utf-8: {err}"
-                        ))
-                    })?
-                );
-            }
-            let _ = writeln!(&mut out, "-----END {block_label}-----");
-            Ok(out.into_bytes())
-        }
-        MaterializeKind::SshPublicKeyShape => {
-            let seed = Seed::from_text(&spec.seed);
-            let mut bytes = [0u8; 32];
-            seed.fill_bytes(&mut bytes);
-            Ok(format!(
-                "ssh-ed25519 {} {}\n",
-                BASE64_STD.encode(bytes),
-                normalize_ssh_comment(&label)
-            )
-            .into_bytes())
-        }
-        MaterializeKind::TokenApiKey => Ok(fx
-            .token(&label, TokenSpec::api_key())
-            .value()
-            .as_bytes()
-            .to_vec()),
-    }
-}
-
 fn verify_fixture_bytes(path: &Path, expected: &[u8]) -> Result<(), MaterializeError> {
     let actual = fs::read(path)?;
     if actual != expected {
@@ -536,42 +442,6 @@ fn fallback_label(path: &Path) -> String {
             }
         })
         .collect()
-}
-
-fn normalize_pem_label(label: &str) -> String {
-    let normalized: String = label
-        .chars()
-        .map(|ch| {
-            if ch.is_ascii_alphanumeric() {
-                ch.to_ascii_uppercase()
-            } else {
-                '_'
-            }
-        })
-        .collect();
-    if normalized.is_empty() {
-        "SECRET".to_string()
-    } else {
-        normalized
-    }
-}
-
-fn normalize_ssh_comment(label: &str) -> String {
-    let normalized: String = label
-        .chars()
-        .map(|ch| {
-            if ch.is_ascii_alphanumeric() || ch == '.' || ch == '_' || ch == '-' {
-                ch
-            } else {
-                '-'
-            }
-        })
-        .collect();
-    if normalized.is_empty() {
-        "fixture".to_string()
-    } else {
-        normalized
-    }
 }
 
 fn fixture_const_name(spec: &MaterializeFixtureSpec) -> String {
